TECH-534. Fix broken links within the tutorials.
When we rename notebooks it also causes references to become deadlinks.
Ideally we'll be able to proactively catch this in the future, but for
next in this commit we just fix the occurrences of broken links. I found
these by grepping for `/tutorials/` and then checking each occurrence
(there aren't that many of them).