TECH-559. Verify that the notebooks can build when opening a PR
We would like to see feedback on changes to this repo's notebooks.
Currently if a change to this repo "breaks" the building of the docs we
won't find out until after merging. This is a longer feedback cycle than
we desire.

This commit introduces an argo workflow that is similar to the workflow
we use for building and publishing the docs, except we stop short of
publishing the docs.