TECH-797. Add paragraph about password prompt to auth docs
We were talking with a user who was (reasonably) confused about where to
enter the username and password that we give them. This commit adds a
paragraph which add a little more detail about the behaviour of the
py-spatiafi package.

Note that we do mention the file for credentials in the paragraph above,
but we don't mention prompting for password.